🛡️ Sentinel: [HIGH] Fix XSS vulnerabilities in user-provided URLs
Abstracts the `sanitizeUrl` function to `src/utils/sanitizeUrl.ts` and applies it globally to user-provided URLs in `href` properties across components to prevent Cross-Site Scripting (XSS).
